Move pointer_grab_start_data into grab
