+++ src/grab/resize_grab.rs
@@ -431,7 +431,7 @@ pub fn resize_request_client(
 ) {
     let pointer = seat.get_pointer().expect("seat had no pointer");
 
-    if let Some(start_data) = crate::pointer::pointer_grab_start_data(&pointer, surface, serial) {
+    if let Some(start_data) = crate::grab::pointer_grab_start_data(&pointer, surface, serial) {
         let Some(window) = state.window_for_surface(surface) else {
             tracing::error!("Surface had no window, cancelling resize request");
             return;
